+++ openspec/changes/agent-codex-split-cli-main-args-dispatch-2026-04-22-13-48/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-22
