+++ src/cli/main.js
@@ -8,6 +8,40 @@ const hooksModule = require('../hooks');
 const sandboxModule = require('../sandbox');
 const toolchainModule = require('../toolchain');
 const finishModule = require('../finish');
+const {
+  gitRun,
+  resolveRepoRoot,
+  isGitRepo,
+  discoverNestedGitRepos,
+} = require('../git');
+const {
+  run,
+  extractTargetedArgs,
+  packageAssetEnv,
+  runPackageAsset,
+  runReviewBotCommand,
+  invokePackageAsset,
+} = require('../core/runtime');
+const {
+  normalizeManagedForcePath,
+  parseCommonArgs,
+  parseSetupArgs,
+  parseDoctorArgs,
+  parseTargetFlag,
+  parseReviewArgs,
+  parseAgentsArgs,
+  parseReportArgs,
+  parseSyncArgs,
+  parseCleanupArgs,
+  parseMergeArgs,
+  parseFinishArgs,
+} = require('./args');
+const {
+  maybeSuggestCommand,
+  normalizeCommandOrThrow,
+  warnDeprecatedAlias,
+  extractFlag,
+} = require('./dispatch');
 
 let sandboxApi;
 let toolchainApi;
@@ -93,7 +127,6 @@ const GUARDEX_REPO_TOGGLE_ENV = 'GUARDEX_ON';
 const DEFAULT_PROTECTED_BRANCHES = ['dev', 'main', 'master'];
 const DEFAULT_BASE_BRANCH = 'dev';
 const DEFAULT_SYNC_STRATEGY = 'rebase';
-const DEFAULT_SHADOW_CLEANUP_IDLE_MINUTES = 60;
 const COMPOSE_HINT_FILES = [
   'docker-compose.yml',
   'docker-compose.yaml',
@@ -269,13 +302,6 @@ const OMX_SCAFFOLD_DIRECTORIES = [
 const OMX_SCAFFOLD_FILES = new Map([
   ['.omx/notepad.md', '\n\n## WORKING MEMORY\n'],
   ['.omx/project-memory.json', '{}\n'],
-]);
-const TARGETED_FORCEABLE_MANAGED_PATHS = new Set([
-  'AGENTS.md',
-  '.gitignore',
-  ...Array.from(OMX_SCAFFOLD_FILES.keys()),
-  ...REQUIRED_MANAGED_REPO_FILES,
-  ...LEGACY_WORKFLOW_SHIMS,
 ]);
 const COMMAND_TYPO_ALIASES = new Map([
   ['relaese', 'release'],
@@ -778,100 +804,6 @@ NOTES
   }
 }
 
-function run(cmd, args, options = {}) {
-  return cp.spawnSync(cmd, args, {
-    encoding: 'utf8',
-    stdio: options.stdio || 'pipe',
-    cwd: options.cwd,
-    env: options.env ? { ...process.env, ...options.env } : process.env,
-    timeout: options.timeout,
-  });
-}
-
-function extractTargetedArgs(rawArgs, defaultTarget = process.cwd()) {
-  const passthrough = [];
-  let target = defaultTarget;
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target' || arg === '-t') {
-      target = requireValue(rawArgs, index, '--target');
-      index += 1;
-      continue;
-    }
-    passthrough.push(arg);
-  }
-
-  return { target, passthrough };
-}
-
-function packageAssetEnv(extraEnv = {}) {
-  return {
-    GUARDEX_CLI_ENTRY: __filename,
-    GUARDEX_NODE_BIN: process.execPath,
-    ...extraEnv,
-  };
-}
-
-function packageAssetPath(assetKey) {
-  const assetPath = PACKAGE_SCRIPT_ASSETS[assetKey];
-  if (!assetPath) {
-    throw new Error(`Unknown package asset: ${assetKey}`);
-  }
-  if (!fs.existsSync(assetPath)) {
-    throw new Error(`Missing package asset: ${assetPath}`);
-  }
-  return assetPath;
-}
-
-function runPackageAsset(assetKey, rawArgs, options = {}) {
-  const assetPath = packageAssetPath(assetKey);
-  let cmd = 'bash';
-  if (assetPath.endsWith('.py')) {
-    cmd = 'python3';
-  } else if (assetPath.endsWith('.js')) {
-    cmd = process.execPath;
-  }
-  return run(cmd, [assetPath, ...rawArgs], {
-    cwd: options.cwd || process.cwd(),
-    stdio: options.stdio || 'pipe',
-    timeout: options.timeout,
-    env: packageAssetEnv(options.env),
-  });
-}
-
-function repoLocalLegacyScriptPath(repoRoot, relativePath) {
-  const assetPath = path.join(repoRoot, relativePath);
-  return fs.existsSync(assetPath) ? assetPath : null;
-}
-
-function runReviewBotCommand(repoRoot, rawArgs, options = {}) {
-  const legacyScript = repoLocalLegacyScriptPath(repoRoot, 'scripts/review-bot-watch.sh');
-  if (legacyScript) {
-    return run('bash', [legacyScript, ...rawArgs], {
-      cwd: repoRoot,
-      stdio: options.stdio || 'pipe',
-      timeout: options.timeout,
-      env: packageAssetEnv(options.env),
-    });
-  }
-  return runPackageAsset('reviewBot', rawArgs, {
-    ...options,
-    cwd: repoRoot,
-  });
-}
-
-function invokePackageAsset(assetKey, rawArgs, options = {}) {
-  const result = runPackageAsset(assetKey, rawArgs, options);
-  if (result.stdout) process.stdout.write(result.stdout);
-  if (result.stderr) process.stderr.write(result.stderr);
-  if (result.status !== 0) {
-    throw new Error(`${assetKey} command failed with status ${result.status}`);
-  }
-  process.exitCode = 0;
-  return result;
-}
-
 function formatElapsedDuration(ms) {
   const durationMs = Number.isFinite(ms) ? Math.max(0, ms) : 0;
   if (durationMs < 1000) {
@@ -1021,114 +953,6 @@ function printAutoFinishSummary(summary, options = {}) {
     const detail = verbose ? details[0] : summarizeAutoFinishDetail(details[0]);
     console.log(colorizeDoctorOutput(`[${TOOL_NAME}] ${detail}`, detectAutoFinishDetailStatus(detail)));
   }
-}
-
-function gitRun(repoRoot, args, { allowFailure = false } = {}) {
-  const result = run('git', ['-C', repoRoot, ...args]);
-  if (!allowFailure && result.status !== 0) {
-    throw new Error(`git ${args.join(' ')} failed: ${(result.stderr || '').trim()}`);
-  }
-  return result;
-}
-
-function resolveRepoRoot(targetPath) {
-  const resolvedTarget = path.resolve(targetPath || process.cwd());
-  const result = run('git', ['-C', resolvedTarget, 'rev-parse', '--show-toplevel']);
-  if (result.status !== 0) {
-    const stderr = (result.stderr || '').trim();
-    throw new Error(
-      `Target is not inside a git repository: ${resolvedTarget}${stderr ? `\n${stderr}` : ''}`,
-    );
-  }
-  return result.stdout.trim();
-}
-
-function isGitRepo(targetPath) {
-  const resolvedTarget = path.resolve(targetPath || process.cwd());
-  const result = run('git', ['-C', resolvedTarget, 'rev-parse', '--show-toplevel']);
-  return result.status === 0;
-}
-
-const NESTED_REPO_DEFAULT_MAX_DEPTH = 6;
-const NESTED_REPO_DEFAULT_SKIP_DIRS = new Set([
-  'node_modules',
-  '.git',
-  'dist',
-  'build',
-  '.next',
-  '.cache',
-  'target',
-  'vendor',
-  '.venv',
-  '.pnpm-store',
-]);
-function discoverNestedGitRepos(rootPath, opts = {}) {
-  const maxDepth = Number.isFinite(opts.maxDepth) ? Math.max(1, opts.maxDepth) : NESTED_REPO_DEFAULT_MAX_DEPTH;
-  const extraSkip = new Set(Array.isArray(opts.extraSkip) ? opts.extraSkip : []);
-  const includeSubmodules = Boolean(opts.includeSubmodules);
-  const resolvedRoot = path.resolve(rootPath);
-
-  const rootCommonDir = (() => {
-    const result = run('git', ['-C', resolvedRoot, 'rev-parse', '--git-common-dir'], { cwd: resolvedRoot });
-    if (result.status !== 0) return null;
-    const raw = result.stdout.trim();
-    if (!raw) return null;
-    return path.resolve(resolvedRoot, raw);
-  })();
-
-  const worktreeSkipAbsolutes = AGENT_WORKTREE_RELATIVE_DIRS.map((relativeDir) => path.join(resolvedRoot, relativeDir));
-  const found = new Set();
-  found.add(resolvedRoot);
-
-  function shouldSkipDir(dirName) {
-    return NESTED_REPO_DEFAULT_SKIP_DIRS.has(dirName) || extraSkip.has(dirName);
-  }
-
-  function walk(currentPath, depth) {
-    if (depth > maxDepth) return;
-    let entries;
-    try {
-      entries = fs.readdirSync(currentPath, { withFileTypes: true });
-    } catch {
-      return;
-    }
-
-    for (const entry of entries) {
-      const entryPath = path.join(currentPath, entry.name);
-
-      if (entry.name === '.git') {
-        if (entry.isDirectory()) {
-          if (entryPath === path.join(resolvedRoot, '.git')) continue;
-          found.add(path.dirname(entryPath));
-        } else if (includeSubmodules && entry.isFile()) {
-          found.add(path.dirname(entryPath));
-        }
-        continue;
-      }
-
-      if (!entry.isDirectory() || entry.isSymbolicLink()) continue;
-      if (shouldSkipDir(entry.name)) continue;
-      if (worktreeSkipAbsolutes.includes(entryPath)) continue;
-      walk(entryPath, depth + 1);
-    }
-  }
-
-  walk(resolvedRoot, 0);
-
-  const filtered = Array.from(found).filter((repoPath) => {
-    if (repoPath === resolvedRoot) return true;
-    if (!rootCommonDir) return true;
-    const childResult = run('git', ['-C', repoPath, 'rev-parse', '--git-common-dir'], { cwd: repoPath });
-    if (childResult.status !== 0) return true;
-    const childCommonDirRaw = childResult.stdout.trim();
-    if (!childCommonDirRaw) return true;
-    const childCommonDir = path.resolve(repoPath, childCommonDirRaw);
-    return childCommonDir !== rootCommonDir;
-  });
-
-  const [root, ...rest] = filtered;
-  rest.sort((a, b) => a.localeCompare(b));
-  return [root, ...rest];
 }
 
 function toDestinationPath(relativeTemplatePath) {
@@ -1857,45 +1681,6 @@ function configureHooks(repoRoot, dryRun) {
   return { status: 'set', key: 'core.hooksPath', value: '.githooks' };
 }
 
-function requireValue(rawArgs, index, flagName) {
-  const value = rawArgs[index + 1];
-  if (!value || value.startsWith('-')) {
-    throw new Error(`${flagName} requires a value`);
-  }
-  return value;
-}
-
-function normalizeManagedForcePath(rawPath) {
-  if (typeof rawPath !== 'string') {
-    return null;
-  }
-  const normalized = path.posix.normalize(rawPath.replace(/\\/g, '/'));
-  if (!normalized || normalized === '.' || normalized.startsWith('../') || path.posix.isAbsolute(normalized)) {
-    return null;
-  }
-  return normalized.startsWith('./') ? normalized.slice(2) : normalized;
-}
-
-function collectForceManagedPaths(rawArgs, startIndex) {
-  const forceManagedPaths = [];
-  let nextIndex = startIndex;
-
-  while (nextIndex + 1 < rawArgs.length) {
-    const candidate = rawArgs[nextIndex + 1];
-    if (!candidate || candidate.startsWith('-')) {
-      break;
-    }
-    const normalized = normalizeManagedForcePath(candidate);
-    if (!normalized || !TARGETED_FORCEABLE_MANAGED_PATHS.has(normalized)) {
-      throw new Error(`Unknown managed path after --force: ${candidate}`);
-    }
-    forceManagedPaths.push(normalized);
-    nextIndex += 1;
-  }
-
-  return { forceManagedPaths, nextIndex };
-}
-
 function appendForceArgs(args, options) {
   if (!options.force) {
     return;
@@ -1916,190 +1701,6 @@ function shouldForceManagedPath(options, relativePath) {
   }
   const normalized = normalizeManagedForcePath(relativePath);
   return normalized !== null && targetedPaths.includes(normalized);
-}
-
-function parseCommonArgs(rawArgs, defaults) {
-  const options = { ...defaults };
-  const supportsForce = Object.prototype.hasOwnProperty.call(options, 'force');
-  if (supportsForce && !Array.isArray(options.forceManagedPaths)) {
-    options.forceManagedPaths = [];
-  }
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target' || arg === '-t') {
-      options.target = requireValue(rawArgs, index, '--target');
-      index += 1;
-      continue;
-    }
-    if (arg === '--dry-run') {
-      options.dryRun = true;
-      continue;
-    }
-    if (arg === '--skip-agents') {
-      options.skipAgents = true;
-      continue;
-    }
-    if (arg === '--skip-package-json') {
-      options.skipPackageJson = true;
-      continue;
-    }
-    if (arg === '--force') {
-      if (!supportsForce) {
-        throw new Error(`Unknown option: ${arg}`);
-      }
-      options.force = true;
-      const parsed = collectForceManagedPaths(rawArgs, index);
-      if (parsed.forceManagedPaths.length > 0) {
-        options.forceManagedPaths = Array.from(
-          new Set([...(options.forceManagedPaths || []), ...parsed.forceManagedPaths]),
-        );
-      }
-      index = parsed.nextIndex;
-      continue;
-    }
-    if (arg === '--keep-stale-locks') {
-      options.dropStaleLocks = false;
-      continue;
-    }
-    if (arg === '--json') {
-      options.json = true;
-      continue;
-    }
-    if (arg === '--yes-global-install') {
-      options.yesGlobalInstall = true;
-      continue;
-    }
-    if (arg === '--no-global-install') {
-      options.noGlobalInstall = true;
-      continue;
-    }
-    if (arg === '--no-gitignore') {
-      options.skipGitignore = true;
-      continue;
-    }
-    if (arg === '--allow-protected-base-write') {
-      options.allowProtectedBaseWrite = true;
-      continue;
-    }
-    if (Object.prototype.hasOwnProperty.call(options, 'waitForMerge') && arg === '--wait-for-merge') {
-      options.waitForMerge = true;
-      continue;
-    }
-    if (Object.prototype.hasOwnProperty.call(options, 'waitForMerge') && arg === '--no-wait-for-merge') {
-      options.waitForMerge = false;
-      continue;
-    }
-
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  if (!options.target) {
-    throw new Error('--target requires a path value');
-  }
-
-  return options;
-}
-
-function parseRepoTraversalArgs(rawArgs, defaults) {
-  const traversalDefaults = {
-    ...defaults,
-    recursive: true,
-    nestedMaxDepth: NESTED_REPO_DEFAULT_MAX_DEPTH,
-    nestedSkipDirs: [],
-    includeSubmodules: false,
-  };
-  const forwardedArgs = [];
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--no-recursive' || arg === '--no-nested' || arg === '--single-repo' || arg === '--current') {
-      traversalDefaults.recursive = false;
-      continue;
-    }
-    if (arg === '--recursive' || arg === '--nested') {
-      traversalDefaults.recursive = true;
-      continue;
-    }
-    if (arg === '--max-depth') {
-      const raw = requireValue(rawArgs, index, '--max-depth');
-      const parsed = Number.parseInt(raw, 10);
-      if (!Number.isFinite(parsed) || parsed < 1) {
-        throw new Error('--max-depth requires a positive integer');
-      }
-      traversalDefaults.nestedMaxDepth = parsed;
-      index += 1;
-      continue;
-    }
-    if (arg === '--skip-nested') {
-      const raw = requireValue(rawArgs, index, '--skip-nested');
-      traversalDefaults.nestedSkipDirs.push(raw);
-      index += 1;
-      continue;
-    }
-    if (arg === '--include-submodules') {
-      traversalDefaults.includeSubmodules = true;
-      continue;
-    }
-    forwardedArgs.push(arg);
-  }
-
-  return parseCommonArgs(forwardedArgs, traversalDefaults);
-}
-
-function parseSetupArgs(rawArgs, defaults) {
-  const setupDefaults = {
-    ...defaults,
-    parentWorkspaceView: false,
-  };
-  const forwardedArgs = [];
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--parent-workspace-view') {
-      setupDefaults.parentWorkspaceView = true;
-      continue;
-    }
-    if (arg === '--no-parent-workspace-view') {
-      setupDefaults.parentWorkspaceView = false;
-      continue;
-    }
-    forwardedArgs.push(arg);
-  }
-
-  return parseRepoTraversalArgs(forwardedArgs, setupDefaults);
-}
-
-function parseDoctorArgs(rawArgs) {
-  const doctorDefaults = {
-    target: process.cwd(),
-    force: false,
-    dropStaleLocks: true,
-    skipAgents: false,
-    skipPackageJson: false,
-    skipGitignore: false,
-    dryRun: false,
-    json: false,
-    allowProtectedBaseWrite: false,
-    waitForMerge: true,
-    verboseAutoFinish: false,
-  };
-  const forwardedArgs = [];
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--verbose-auto-finish') {
-      doctorDefaults.verboseAutoFinish = true;
-      continue;
-    }
-    if (arg === '--compact-auto-finish') {
-      doctorDefaults.verboseAutoFinish = false;
-      continue;
-    }
-    forwardedArgs.push(arg);
-  }
-
-  return parseRepoTraversalArgs(forwardedArgs, doctorDefaults);
 }
 
 function normalizeWorkspacePath(relativePath) {
@@ -3399,171 +3000,6 @@ function runSetupInSandbox(options, blocked, repoLabel = '') {
   };
 }
 
-function parseTargetFlag(rawArgs, defaultTarget = process.cwd()) {
-  const remaining = [];
-  let target = defaultTarget;
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--target requires a path value');
-      }
-      target = next;
-      index += 1;
-      continue;
-    }
-    remaining.push(arg);
-  }
-
-  return { target, args: remaining };
-}
-
-function parseReviewArgs(rawArgs) {
-  const parsed = parseTargetFlag(rawArgs, process.cwd());
-  const passthroughArgs = [...parsed.args];
-  if (passthroughArgs[0] === 'start') {
-    passthroughArgs.shift();
-  }
-  return {
-    target: parsed.target,
-    passthroughArgs,
-  };
-}
-
-function parseAgentsArgs(rawArgs) {
-  const parsed = parseTargetFlag(rawArgs, process.cwd());
-  const [subcommandRaw = '', ...rest] = parsed.args;
-  const subcommand = subcommandRaw || 'status';
-  const options = {
-    target: parsed.target,
-    subcommand,
-    reviewIntervalSeconds: 30,
-    cleanupIntervalSeconds: 60,
-    idleMinutes: DEFAULT_SHADOW_CLEANUP_IDLE_MINUTES,
-  };
-
-  for (let index = 0; index < rest.length; index += 1) {
-    const arg = rest[index];
-    if (arg === '--review-interval') {
-      const next = rest[index + 1];
-      if (!next) {
-        throw new Error('--review-interval requires an integer seconds value');
-      }
-      const parsedValue = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsedValue) || parsedValue < 5) {
-        throw new Error('--review-interval must be an integer >= 5 seconds');
-      }
-      options.reviewIntervalSeconds = parsedValue;
-      index += 1;
-      continue;
-    }
-    if (arg === '--cleanup-interval') {
-      const next = rest[index + 1];
-      if (!next) {
-        throw new Error('--cleanup-interval requires an integer seconds value');
-      }
-      const parsedValue = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsedValue) || parsedValue < 5) {
-        throw new Error('--cleanup-interval must be an integer >= 5 seconds');
-      }
-      options.cleanupIntervalSeconds = parsedValue;
-      index += 1;
-      continue;
-    }
-    if (arg === '--idle-minutes') {
-      const next = rest[index + 1];
-      if (!next) {
-        throw new Error('--idle-minutes requires an integer minutes value');
-      }
-      const parsedValue = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsedValue) || parsedValue < 1) {
-        throw new Error('--idle-minutes must be an integer >= 1');
-      }
-      options.idleMinutes = parsedValue;
-      index += 1;
-      continue;
-    }
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  if (!['start', 'stop', 'status'].includes(options.subcommand)) {
-    throw new Error(`Unknown agents subcommand: ${options.subcommand}`);
-  }
-
-  return options;
-}
-
-function parseReportArgs(rawArgs) {
-  const options = {
-    target: process.cwd(),
-    subcommand: '',
-    repo: '',
-    scorecardJson: '',
-    outputDir: '',
-    date: '',
-    dryRun: false,
-    json: false,
-  };
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) throw new Error('--target requires a path value');
-      options.target = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--repo') {
-      const next = rawArgs[index + 1];
-      if (!next) throw new Error('--repo requires a value like github.com/owner/repo');
-      options.repo = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--scorecard-json') {
-      const next = rawArgs[index + 1];
-      if (!next) throw new Error('--scorecard-json requires a path value');
-      options.scorecardJson = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--output-dir') {
-      const next = rawArgs[index + 1];
-      if (!next) throw new Error('--output-dir requires a path value');
-      options.outputDir = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--date') {
-      const next = rawArgs[index + 1];
-      if (!next) throw new Error('--date requires a YYYY-MM-DD value');
-      options.date = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--dry-run') {
-      options.dryRun = true;
-      continue;
-    }
-    if (arg === '--json') {
-      options.json = true;
-      continue;
-    }
-    if (arg.startsWith('-')) {
-      throw new Error(`Unknown option: ${arg}`);
-    }
-    if (!options.subcommand) {
-      options.subcommand = arg;
-      continue;
-    }
-    throw new Error(`Unexpected argument: ${arg}`);
-  }
-
-  return options;
-}
 
 function todayDateStamp() {
   return new Date().toISOString().slice(0, 10);
@@ -4195,401 +3631,6 @@ function lockRegistryStatus(repoRoot) {
   return { dirty: true, untracked };
 }
 
-function parseSyncArgs(rawArgs) {
-  const options = {
-    target: process.cwd(),
-    check: false,
-    base: '',
-    strategy: '',
-    ffOnly: false,
-    dryRun: false,
-    json: false,
-    allAgentBranches: false,
-    allowNonAgent: false,
-    allowDirty: false,
-  };
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--target requires a path value');
-      }
-      options.target = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--base') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--base requires a branch value');
-      }
-      options.base = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--strategy') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--strategy requires a value (rebase|merge)');
-      }
-      options.strategy = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--check') {
-      options.check = true;
-      continue;
-    }
-    if (arg === '--ff-only') {
-      options.ffOnly = true;
-      continue;
-    }
-    if (arg === '--dry-run') {
-      options.dryRun = true;
-      continue;
-    }
-    if (arg === '--json') {
-      options.json = true;
-      continue;
-    }
-    if (arg === '--all-agent-branches') {
-      options.allAgentBranches = true;
-      continue;
-    }
-    if (arg === '--allow-non-agent') {
-      options.allowNonAgent = true;
-      continue;
-    }
-    if (arg === '--allow-dirty') {
-      options.allowDirty = true;
-      continue;
-    }
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  return options;
-}
-
-function parseCleanupArgs(rawArgs) {
-  const options = {
-    target: process.cwd(),
-    base: '',
-    branch: '',
-    dryRun: false,
-    forceDirty: false,
-    keepRemote: false,
-    keepCleanWorktrees: false,
-    includePrMerged: false,
-    idleMinutes: 0,
-    watch: false,
-    intervalSeconds: 60,
-    once: false,
-    maxBranches: 0,
-  };
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--target requires a path value');
-      }
-      options.target = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--base') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--base requires a branch value');
-      }
-      options.base = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--branch') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--branch requires an agent branch value');
-      }
-      options.branch = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--dry-run') {
-      options.dryRun = true;
-      continue;
-    }
-    if (arg === '--force-dirty') {
-      options.forceDirty = true;
-      continue;
-    }
-    if (arg === '--keep-remote') {
-      options.keepRemote = true;
-      continue;
-    }
-    if (arg === '--keep-clean-worktrees') {
-      options.keepCleanWorktrees = true;
-      continue;
-    }
-    if (arg === '--include-pr-merged') {
-      options.includePrMerged = true;
-      continue;
-    }
-    if (arg === '--idle-minutes') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--idle-minutes requires an integer value');
-      }
-      const parsed = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsed) || parsed < 0) {
-        throw new Error('--idle-minutes must be an integer >= 0');
-      }
-      options.idleMinutes = parsed;
-      index += 1;
-      continue;
-    }
-    if (arg === '--watch') {
-      options.watch = true;
-      continue;
-    }
-    if (arg === '--interval') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--interval requires an integer seconds value');
-      }
-      const parsed = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsed) || parsed < 5) {
-        throw new Error('--interval must be an integer >= 5 seconds');
-      }
-      options.intervalSeconds = parsed;
-      index += 1;
-      continue;
-    }
-    if (arg === '--once') {
-      options.once = true;
-      continue;
-    }
-    if (arg === '--max-branches') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--max-branches requires an integer value');
-      }
-      const parsed = Number.parseInt(next, 10);
-      if (!Number.isInteger(parsed) || parsed < 1) {
-        throw new Error('--max-branches must be an integer >= 1');
-      }
-      options.maxBranches = parsed;
-      index += 1;
-      continue;
-    }
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  if (options.watch && options.idleMinutes === 0) {
-    options.idleMinutes = DEFAULT_SHADOW_CLEANUP_IDLE_MINUTES;
-  }
-
-  return options;
-}
-
-function parseMergeArgs(rawArgs) {
-  const options = {
-    target: process.cwd(),
-    base: '',
-    into: '',
-    branches: [],
-    task: '',
-    agent: '',
-  };
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--target requires a path value');
-      }
-      options.target = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--base') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--base requires a branch value');
-      }
-      options.base = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--into') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--into requires an agent/* branch value');
-      }
-      options.into = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--branch') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--branch requires an agent/* branch value');
-      }
-      options.branches.push(next);
-      index += 1;
-      continue;
-    }
-    if (arg === '--task') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--task requires a task value');
-      }
-      options.task = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--agent') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--agent requires an agent value');
-      }
-      options.agent = next;
-      index += 1;
-      continue;
-    }
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  if (options.branches.length === 0) {
-    throw new Error('merge requires at least one --branch <agent/*> input');
-  }
-
-  return options;
-}
-
-function parseFinishArgs(rawArgs, defaults = {}) {
-  const options = {
-    target: process.cwd(),
-    base: '',
-    branch: '',
-    all: false,
-    dryRun: false,
-    waitForMerge: defaults.waitForMerge ?? true,
-    cleanup: defaults.cleanup ?? true,
-    keepRemote: false,
-    noAutoCommit: false,
-    failFast: false,
-    commitMessage: '',
-    mergeMode: defaults.mergeMode || 'pr',
-  };
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const arg = rawArgs[index];
-    if (arg === '--target') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--target requires a path value');
-      }
-      options.target = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--base') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--base requires a branch value');
-      }
-      options.base = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--branch') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--branch requires an agent/* branch value');
-      }
-      options.branch = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--commit-message') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--commit-message requires a value');
-      }
-      options.commitMessage = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--all') {
-      options.all = true;
-      continue;
-    }
-    if (arg === '--dry-run') {
-      options.dryRun = true;
-      continue;
-    }
-    if (arg === '--wait-for-merge') {
-      options.waitForMerge = true;
-      continue;
-    }
-    if (arg === '--no-wait-for-merge') {
-      options.waitForMerge = false;
-      continue;
-    }
-    if (arg === '--via-pr') {
-      options.mergeMode = 'pr';
-      continue;
-    }
-    if (arg === '--direct-only') {
-      options.mergeMode = 'direct';
-      continue;
-    }
-    if (arg === '--mode') {
-      const next = rawArgs[index + 1];
-      if (!next) {
-        throw new Error('--mode requires a value');
-      }
-      if (!['auto', 'direct', 'pr'].includes(next)) {
-        throw new Error(`Invalid --mode value: ${next} (expected auto|direct|pr)`);
-      }
-      options.mergeMode = next;
-      index += 1;
-      continue;
-    }
-    if (arg === '--cleanup') {
-      options.cleanup = true;
-      continue;
-    }
-    if (arg === '--no-cleanup') {
-      options.cleanup = false;
-      continue;
-    }
-    if (arg === '--keep-remote') {
-      options.keepRemote = true;
-      continue;
-    }
-    if (arg === '--no-auto-commit') {
-      options.noAutoCommit = true;
-      continue;
-    }
-    if (arg === '--fail-fast') {
-      options.failFast = true;
-      continue;
-    }
-    throw new Error(`Unknown option: ${arg}`);
-  }
-
-  if (options.branch && !options.branch.startsWith('agent/')) {
-    throw new Error(`--branch must reference an agent/* branch (received: ${options.branch})`);
-  }
-
-  return options;
-}
 
 function listAgentWorktrees(repoRoot) {
   const result = gitRun(repoRoot, ['worktree', 'list', '--porcelain'], { allowFailure: true });
@@ -6046,10 +5087,11 @@ function doctor(rawArgs) {
   const topRepoRoot = resolveRepoRoot(options.target);
   const discoveredRepos = options.recursive
     ? discoverNestedGitRepos(topRepoRoot, {
-      maxDepth: options.nestedMaxDepth,
-      extraSkip: options.nestedSkipDirs,
-      includeSubmodules: options.includeSubmodules,
-    })
+        maxDepth: options.nestedMaxDepth,
+        extraSkip: options.nestedSkipDirs,
+        includeSubmodules: options.includeSubmodules,
+        skipRelativeDirs: AGENT_WORKTREE_RELATIVE_DIRS,
+      })
     : [topRepoRoot];
 
   if (discoveredRepos.length > 1) {
@@ -6680,6 +5722,7 @@ function setup(rawArgs) {
         maxDepth: options.nestedMaxDepth,
         extraSkip: options.nestedSkipDirs,
         includeSubmodules: options.includeSubmodules,
+        skipRelativeDirs: AGENT_WORKTREE_RELATIVE_DIRS,
       })
     : [topRepoRoot];
 
@@ -7550,78 +6593,6 @@ function protect(rawArgs) {
   }
 
   throw new Error(`Unknown protect subcommand: ${subcommand}`);
-}
-
-function levenshteinDistance(a, b) {
-  const rows = a.length + 1;
-  const cols = b.length + 1;
-  const matrix = Array.from({ length: rows }, () => Array(cols).fill(0));
-
-  for (let i = 0; i < rows; i += 1) matrix[i][0] = i;
-  for (let j = 0; j < cols; j += 1) matrix[0][j] = j;
-
-  for (let i = 1; i < rows; i += 1) {
-    for (let j = 1; j < cols; j += 1) {
-      const cost = a[i - 1] === b[j - 1] ? 0 : 1;
-      matrix[i][j] = Math.min(
-        matrix[i - 1][j] + 1, // deletion
-        matrix[i][j - 1] + 1, // insertion
-        matrix[i - 1][j - 1] + cost, // substitution
-      );
-    }
-  }
-  return matrix[a.length][b.length];
-}
-
-function maybeSuggestCommand(command) {
-  let best = null;
-  let bestDistance = Number.POSITIVE_INFINITY;
-
-  for (const candidate of SUGGESTIBLE_COMMANDS) {
-    const dist = levenshteinDistance(command, candidate);
-    if (dist < bestDistance) {
-      bestDistance = dist;
-      best = candidate;
-    }
-  }
-
-  if (best && bestDistance <= 2) {
-    return best;
-  }
-
-  return null;
-}
-
-function normalizeCommandOrThrow(command) {
-  if (COMMAND_TYPO_ALIASES.has(command)) {
-    const mapped = COMMAND_TYPO_ALIASES.get(command);
-    console.log(`[${TOOL_NAME}] Interpreting '${command}' as '${mapped}'.`);
-    return mapped;
-  }
-  return command;
-}
-
-function warnDeprecatedAlias(aliasName) {
-  const entry = DEPRECATED_COMMAND_ALIASES.get(aliasName);
-  if (!entry) return;
-  console.error(
-    `[${TOOL_NAME}] '${aliasName}' is deprecated and will be removed in a future major release. ` +
-    `Use: ${entry.hint}`,
-  );
-}
-
-function extractFlag(args, ...names) {
-  const flagSet = new Set(names);
-  let found = false;
-  const remaining = [];
-  for (const arg of args) {
-    if (flagSet.has(arg)) {
-      found = true;
-    } else {
-      remaining.push(arg);
-    }
-  }
-  return { found, remaining };
 }
 
 function main() {
